projects/dataflow-framework/pyproject.toml
[project]
name = "dataflow-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "typer>=0.17.3",
]
